flask_app/requirements.txt
Flask
mysql.connector
requests
tensorflow-cpu
apscheduler
python-dotenv